+++ packages/util/src/u8a/eq.ts
@@ -1,32 +1,23 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { u8aCmp } from './cmp';
 import { u8aToU8a } from './toU8a';
 
 function equals (a: Uint8Array, b: Uint8Array): boolean {
-  if (a.length !== b.length) {
-    return false;
-  }
-
-  for (let i = 0; i < a.length; i++) {
-    if (a[i] !== b[i]) {
-      return false;
-    }
-  }
-
-  return true;
+  return (a.length === b.length) && (u8aCmp(a, b) === 0);
 }
 
 /**
  * @name u8aEq
- * @summary Compares two Uint8Arrays.
+ * @summary Compares two Uint8Arrays for equality.
  * @description
  * For `UInt8Array` (or hex string) input values true if there is a match.
  * @example
  * <BR>
  *
  * ```javascript
- * import { { u8aEq } from '@polkadot/util';
+ * import { u8aEq } from '@polkadot/util';
  *
  * u8aEq(new Uint8Array([0x68, 0x65]), new Uint8Array([0x68, 0x65])); // true
  * ```