+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Added `u8aCmp(a, b)` to do sorting compares
+
+
 ## 6.8.1 Jun 14, 2021
 
 Contributed: